+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 * Added rewind and forward buttons for android player.
 * Fixed player UI's jank
 * Added enableSkips and skipsTimeInMilliseconds in BetterPlayerControlsConfiguration
-
+* Changed middle play button behavior (now it's only used for restart player).
 
 ## 0.0.29+1
 * Updated readme